feat: Respawn the car to the previous checkpoint it crossed over.

+++ game/src/game/components/Controllers/PlayerController.h
@@ -36,7 +36,7 @@ class PlayerController final : public Component,
     VehicleCommand command_;
 
     // to respawn the car
-    float button_down_respawn_timer = 0.0f;
+    double button_down_respawn_timer = 0.0f;
 
     void UpdatePowerupControls(const Timestep& delta_time);
     void UpdateCarControls(const Timestep& delta_time);
